split ABOUT.js into 3 links/pages
