--- src/components/general/newbieuniversity/NewbieUniversity.js
@@ -1,12 +1,0 @@
-/* Imported stylesheet */
-import classes from "./NewbieUniversity.module.css";
-
-const Newbie = () => {
-  return (
-    <div className={classes.newbie}>
-      <h1>Showing: Newbie University</h1>
-    </div>
-  );
-};
-
-export default Newbie;
