--- src/components/general/newbieuniversity/NewbieUniversity.module.css
@@ -1,10 +1,0 @@
-.newbie {
-  display: flex;
-  width: 100%;
-  justify-content: center;
-}
-
-.newbie h1 {
-  text-align: center;
-  text-decoration: underline;
-}
