+++ src/components/general/team/Team.module.css
@@ -1,10 +1,21 @@
-.team {
+.teamContainer {
+  display: flex;
+  justify-content: center;
+  padding-bottom: 60px;
+}
+
+.teamContent {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  width: 80%;
   border: solid #ffde59 2px;
   border-radius: 10px;
   padding: 25px 50px;
+  margin: 25px 0px;
 }
 
-.team h2 {
+.teamContent h2 {
   text-align: center;
   margin-bottom: 25px;
   font-weight: 800;
