+++ src/components/general/team/teamcard/TeamCard.js
@@ -1,20 +1,15 @@
 /* Imported stylesheet */
 import classes from "./TeamCard.module.css";
+import imgSrc from "../../../../../assets/staff_logos/fmc.png";
 
-const TeamCard = ({
-  title,
-  imageSrc,
-  imageAlt,
-  discordUsername,
-  twitterUsername,
-}) => {
+const TeamCard = ({ title, imageAlt, discordUsername, twitterUsername }) => {
   return (
     <div className={classes.teamCardContainer}>
       <div className={classes.titleContainer}>
         <h3>{title}</h3>
       </div>
       <div className={classes.infoContainer}>
-        <img src={imageSrc} alt={imageAlt}></img>
+        <img src={imgSrc} alt={imageAlt}></img>
         <h5>
           Discord: <span>{discordUsername}</span>
         </h5>
