+++ src/components/general/whitepaper/Whitepaper.module.css
@@ -1,29 +1,39 @@
-.whitepaper {
+.whitepaperContainer {
+  display: flex;
+  justify-content: center;
+  padding-bottom: 60px;
+}
+
+.whitepaperContent {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  width: 80%;
   border: solid #ffde59 2px;
   border-radius: 10px;
   padding: 25px 50px;
   margin: 25px 0px;
 }
 
-.whitepaper h2 {
+.whitepaperContent h2 {
   text-align: center;
   margin-bottom: 25px;
   font-weight: 800;
   color: rgb(109, 204, 162, 1);
 }
 
-.whitepaper h3 {
+.whitepaperContent h3 {
   text-align: center;
   margin-bottom: 5px;
 }
 
-.whitepaper p {
+.whitepaperContent p {
   text-align: center;
   margin-bottom: 25px;
   line-height: 1.5rem;
 }
 
-.whitepaper button {
+.whitepaperContent button {
   display: flex;
   width: auto;
   margin: auto;
